+++ packages/utils/src/useDeepCompareMemo.ts
@@ -1,7 +1,8 @@
 'use client';
 
 import { useRef } from 'react';
-import isEqual from 'lodash/isEqual';
+
+import deepEquals from './deepEquals';
 
 /** Hook that stores and returns a `T` value. If `newValue` is the same as the stored one, then the stored one is
  * returned to avoid having a component rerender due it being a different object. Otherwise, the `newValue` is stored
@@ -12,7 +13,7 @@ import isEqual from 'lodash/isEqual';
  */
 export default function useDeepCompareMemo<T = unknown>(newValue: T): T {
   const valueRef = useRef<T>(newValue);
-  if (!isEqual(newValue, valueRef.current)) {
+  if (!deepEquals(newValue, valueRef.current)) {
     valueRef.current = newValue;
   }
   return valueRef.current;